Fixed typo in GPUMatrix.cu

+++ Source/Math/GPUMatrix.cu
@@ -2374,7 +2374,7 @@ ElemType GPUMatrix<ElemType>::Max() const
         int resInd = 0;
         cublasIdamax(cuHandle, (CUDA_LONG) GetNumElements(), reinterpret_cast<double*>(Data()), 1, &resInd);
         resInd--;
-        CUDA_CALL(cudaMemcpy(reinterpret_cast<double*>(&res), Data()+ resInd, sizeof(float), cudaMemcpyDeviceToHost));
+        CUDA_CALL(cudaMemcpy(reinterpret_cast<double*>(&res), Data()+ resInd, sizeof(double), cudaMemcpyDeviceToHost));
         return res;
     }
 }
